requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.5
django-classy-tags==3.0.1
django-formtools==2.3
django-nested-admin==3.4.0
django-polymorphic==3.1.0
django-sekizai==3.0.1
django-treebeard==4.5.1
djangocms-admin-style==3.1.1
gitdb==4.0.9
GitPython==3.1.27
Pillow==9.1.0
python-monkey-business==1.0.0
pytz==2022.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
